nir/linker: add some nir-based cross stage uniform validation
For now it only checks that the type of a given uniform on the stage
being processed is the same that the same uniform on a previous stage
already processed.

Note that in opposite to GLSL, that does a similar thing with IR
variables, we are checking that the uniforms are the same based on the
location, as we are basing the ARB_gl_spirv linking on the location.

We are also avoiding name-checks. As they are just debug info, nothing
prevents to have one SPIR-V module with the names and other without
it. It also seems reasonable to think that it is not even needed to
keep the same names on different modules/stages.

As mentioned on the comment included with this patch, on GLSL linker
there are some full cross-stage validation methods, taking into
account more than just the uniforms. It is likely that this gets moved
later to a bigger method like that.

NOTE: needed to get the following (removed) tests passing:
   * tests/spec/arb_gl_spirv/linker/uniform/different-uniform-array-size.shader_test
   * tests/spec/arb_gl_spirv/linker/uniform/different-uniform-type.shader_test

v2: compare two nir variables, instead of compare a nir variable and
    an gl_uniform_storage derived from a nir variable.