mesa/main: fix find_binding_offset for arrays of blocks
For arrays, or arrays of arrays of blocks, it is usual that resources
are only added for the block with base index. But most properties for
members of each index are basically inherited from the base index,
including a uniform being active or not. So we use
linearized_array_index to get the block with the base index.