nir: allow stitching of non-empty block
When stitching two blocks A and B, where A's last instruction is a jump,
it is not required that B is empty; it can be plainly removed.

This can happen in a situation like this:

vec1 1 ssa_1 = load_const (true)
vec1 1 ssa_2 = load_const (false)
block block_1:
[...]
loop {
  vec1 ssa_3 = phi block_1: ssa_2, block_4: ssa_1
  if ssa_3 {
    block block_2:
    [...]
    break
  } else {
    block block_3:
  }
  vec1 ssa_4 = <alu operation>
  if ssa_4 {
    block block_4:
    continue
  } else {
    block block_5:
  }
  block block_6:
  [...]
}

And opt_peel_loop_initial_if is applied. In this case, we would be
ending up stitching block_2 (which finalizes with a jump) with
block_4, which is not empty.

CC: Jason Ekstrand <jason@jlekstrand.net>