intel/compiler: implement more algebraic optimizations
If we want to constant propagate immediates more aggressively to
other 2-src instructions (MUL, ADD, etc) we need to prepare for
the fact that we can end up with 2 immediate sources on these
instructions, or an immediate on the first source. These scenarios
can be triggered by the backend (so after NIR translation), for
example, when we emit code to compute load/store offsets. We should
also support more types than :F.